Excise health from firehose
