+++ .github/workflows/health_internal.yaml
@@ -9,4 +9,4 @@ jobs:
   health:
     uses: ./.github/workflows/health.yaml
     with:
-      checks: "changelog,license"
+      checks: "changelog license"
