+++ pkgs/firehose/bin/firehose.dart
@@ -20,11 +20,10 @@ void main(List<String> arguments) async {
 
     var validate = argResults['validate'] == true;
     var publish = argResults['publish'] == true;
-    var health = argResults['health'] != null;
 
-    if (!validate && !publish && !health) {
-      _usage(argParser, error: '''
-Error: one of --validate, --publish, or --health must be specified.''');
+    if (!validate && !publish) {
+      _usage(argParser,
+          error: 'Error: one of --validate or --publish must be specified.');
       exit(1);
     }
 
@@ -42,8 +41,6 @@ Error: one of --validate, --publish, or --health must be specified.''');
       await firehose.validate();
     } else if (publish) {
       await firehose.publish();
-    } else if (health) {
-      await firehose.healthCheck(argResults['health'] as List);
     }
   } on ArgParserException catch (e) {
     _usage(argParser, error: e.message);
@@ -75,11 +72,6 @@ ArgParser _createArgs() {
       negatable: false,
       help: 'Validate packages and indicate whether --publish would publish '
           'anything.',
-    )
-    ..addMultiOption(
-      'health',
-      defaultsTo: ['version', 'license', 'changelog'],
-      help: 'Check PR health.',
     )
     ..addFlag(
       'publish',
